ingestion/requirements.txt
requests
pandas
pyarrow
python-dotenv
databricks-sql-connector
python-dateutil
certifi
resend
